+++ GitCommands/packages.config
@@ -1,4 +1,6 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <packages>
+  <package id="ConEmu.Console" version="1.0.151207.0" targetFramework="net40-client" />
+  <package id="ConEmu.Control.WinForms" version="1.0.151207.0" targetFramework="net40-client" />
   <package id="SmartFormat.NET" version="1.6.1.0" targetFramework="net40-Client" />
 </packages>
