+++ GitUI/GitUI.csproj
@@ -232,6 +232,9 @@
       <DependentUpon>CommitSummaryUserControl.cs</DependentUpon>
     </Compile>
     <Compile Include="CommandsDialogs\CommitDialog\CommitTemplateItem.cs" />
+    <Compile Include="UserControls\ConsoleProcessControl.cs">
+      <SubType>Component</SubType>
+    </Compile>
     <Compile Include="UserControls\ExListView.cs">
       <SubType>Component</SubType>
     </Compile>
